Use `Arc<dyn Context>` instead of a ref or box
Updates `create_context`, `CryptoFactory`, and `ContextSigner` to
exclusively use an `Arc<dyn Context>` when dealing with a context. This
provides the same functionality as both the boxed and referenced
contexts with respect to ownership and sharing of the context, but it
eliminates the lifetime requirements associated with the plain
reference.

Signed-off-by: Logan Seeley <seeley@bitwise.io>